Eliminated compilation warning.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@21131 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/video/render3d/shader/shaderplugins/glshader_ps1/ps1_emu_nv.h
@@ -93,7 +93,7 @@ public:
   csShaderGLPS1_NV (csGLShader_PS1* shaderPlug)
     : csShaderGLPS1_Common(shaderPlug) 
   {
-    tex_program_num = ~0;
+    tex_program_num = (GLuint)~0;
   }
   virtual ~csShaderGLPS1_NV ()
   {
